[#94] RainbowRecordViewController 네트워크 통신
